floatをdoubleに

+++ src/mmd/PmxData.pyx
@@ -520,7 +520,7 @@ cdef class Sphere(OBB):
     cdef tuple get_collistion(self, MVector3D point, MVector3D root_global_pos, float max_length):
         cdef MMatrix4x4 arm_matrix
         cdef bint collision, near_collision
-        cdef float d, sin_x_theta, sin_y_theta, sin_z_theta, x, x_theta, y, y_theta, z, z_theta, x_distance, z_distance, new_y
+        cdef double d, sin_x_theta, sin_y_theta, sin_z_theta, x, x_theta, y, y_theta, z, z_theta, x_distance, z_distance, new_y
         cdef MVector3D local_point, new_x_local, new_z_local, rep_x_collision_vec, rep_z_collision_vec, x_arm_local, z_arm_local
 
         # 原点との距離が半径未満なら衝突
@@ -601,7 +601,7 @@ cdef class Box(OBB):
     cdef tuple get_collistion(self, MVector3D point, MVector3D root_global_pos, float max_length):
         cdef MMatrix4x4 arm_matrix
         cdef bint collision, near_collision, res1, res2, res3
-        cdef float d, sin_x_theta, sin_y_theta, sin_z_theta, x_theta, y, y_theta, z_theta, new_y, size1, size2, size3, x, z, x_diff, z_diff, x_distance, z_distance
+        cdef double d, sin_x_theta, sin_y_theta, sin_z_theta, x_theta, y, y_theta, z_theta, new_y, size1, size2, size3, x, z, x_diff, z_diff, x_distance, z_distance
         cdef MVector3D b1,  b2, b4, d1, d2, d3, dir1, dir2, dir3, dir_vec, local_point, new_x_local, new_z_local, rep_x_collision_vec, rep_z_collision_vec, t1
         cdef MVector3D x_arm_local, x_base, z_arm_local, z_base
 
@@ -748,10 +748,10 @@ cdef class Capsule(OBB):
     cdef tuple get_collistion(self, MVector3D point, MVector3D root_global_pos, float max_length):
         cdef MMatrix4x4 arm_matrix
         cdef bint collision, near_collision
-        cdef float d, sin_x_theta, sin_y_theta, sin_z_theta, x, x_theta, y, y_theta, z, z_theta, x_distance, z_distance, new_y
+        cdef double d, sin_x_theta, sin_y_theta, sin_z_theta, x, x_theta, y, y_theta, z, z_theta, x_distance, z_distance, new_y
         cdef MVector3D local_point, new_x_local, new_z_local, rep_x_collision_vec, rep_z_collision_vec, x_arm_local, z_arm_local
         cdef MVector3D b1, t1, h, v
-        cdef float ba, bb, bc, ta, tb, tc, lensq, t
+        cdef double ba, bb, bc, ta, tb, tc, lensq, t
 
         # 下辺
         b1 = self.rotated_matrix * MVector3D(0, -self.shape_size.y(), 0)
